Fix Windows build after 273800@main https://bugs.webkit.org/show_bug.cgi?id=268436 rdar://problem/121983146
Reviewed by Ross Kirsling.

Windows build was broken by https://commits.webkit.org/273800@main
because apparently MSVC could not implicitly convert 0ul to size_t.

* Source/WebCore/layout/formattingContexts/block/BlockLayoutState.h:
(WebCore::Layout::BlockLayoutState::LineClamp::allowedLineCount const):

Canonical link: https://commits.webkit.org/273808@main

+++ Source/WebCore/layout/formattingContexts/block/BlockLayoutState.h
@@ -38,7 +38,7 @@ class BlockFormattingContext;
 class BlockLayoutState {
 public:
     struct LineClamp {
-        size_t allowedLineCount() const { return std::max(maximumLineCount - currentLineCount, 0lu); }
+        size_t allowedLineCount() const { return std::max(maximumLineCount - currentLineCount, static_cast<size_t>(0)); }
         size_t maximumLineCount { 0 };
         size_t currentLineCount { 0 };
     };
